project re-structure
