--- .idea/runConfigurations/sshd.xml
@@ -1,11 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="sshd" type="GoApplicationRunConfiguration" factoryName="Go Application" singleton="false">
-    <module name="riotpot" />
-    <working_directory value="$PROJECT_DIR$/" />
-    <kind value="FILE" />
-    <filePath value="$PROJECT_DIR$/src/simulator/sshd.go" />
-    <directory value="$PROJECT_DIR$/" />
-    <output_directory value="$PROJECT_DIR$/bin/" />
-    <method v="2" />
-  </configuration>
-</component>
