+++ LICENSE.md
@@ -1,0 +1,1 @@
+Hacker license!
