+++ Makefile
@@ -1,0 +1,1 @@
+# note: call scripts from /scripts
