+++ api/README.md
@@ -1,0 +1,8 @@
+# `/api`
+
+OpenAPI/Swagger specs, JSON schema files, protocol definition files.
+
+Examples:
+
+* https://github.com/kubernetes/kubernetes/tree/master/api
+* https://github.com/moby/moby/tree/master/api
