+++ assets/README.md
@@ -1,0 +1,3 @@
+# `/assets`
+
+Other assets to go along with your repository (images, logos, etc).
