+++ cmd/riotpot/go.mod
@@ -1,0 +1,3 @@
+module riotpot/cmd/riotpot
+
+go 1.14
