+++ configs/README.md
@@ -1,0 +1,5 @@
+# `/configs`
+
+Configuration file templates or default configs.
+
+Put your `confd` or `consul-template` template files here.
