+++ githooks/README.md
@@ -1,0 +1,3 @@
+# `/githooks`
+
+Git hooks.
