+++ go.mod
@@ -1,0 +1,12 @@
+module github.com/YOUR-USER-OR-ORG-NAME/YOUR-REPO-NAME
+
+go 1.14
+
+require (
+	github.com/anmitsu/go-shlex v0.0.0-20200514113438-38f4b401e2be // indirect
+	github.com/gliderlabs/ssh v0.3.0 // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/kr/pty v1.1.8 // indirect
+	github.com/praetorian-inc/trudy v0.0.0-20170323143650-619d25bba282 // indirect
+	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
+)
