+++ init/README.md
@@ -1,0 +1,3 @@
+# `/init`
+
+System init (systemd, upstart, sysv) and process manager/supervisor (runit, supervisord) configs.
