+++ web/README.md
@@ -1,0 +1,3 @@
+# `/web`
+
+Web application specific components: static web assets, server side templates and SPAs.
