+++ website/README.md
@@ -1,0 +1,8 @@
+# `/website`
+
+This is the place to put your project's website data if you are not using GitHub pages.
+
+Examples:
+
+* https://github.com/hashicorp/vault/tree/master/website
+* https://github.com/perkeep/perkeep/tree/master/website
